Naprawienie cyklicznego importowania
